+++ .gitignore
@@ -4,3 +4,5 @@ node_modules/
 .vscode
 dist
 tmp
+.DS_Store
+packages/.DS_Store
